chore: Update .gitignore, config.toml, and archetypes/default.md
This commit updates the .gitignore file to exclude the /public and /themes directories, as well as the .DS_Store and .hugo_build.lock files.

In the config.toml file, the baseURL is changed to "https://hugo-paper.vercel.app/", the languageCode is set to "ja", and the DefaultContentLanguage is also set to "ja". Additionally, the enableInlineShortcodes parameter is set to true to enable inline shortcodes.

The archetypes/default.md file is modified to include a description field and a tags field with a default value of "tag".

These changes were made to improve the project's configuration and ensure compatibility with the Hugo framework.