chore: Update Minio image URL in blog post
This commit updates the Minio image URL in the "trouble-with-minio.md" blog post. The previous URL pointed to the LAN address, which prevented the images from being displayed outside the local network. The new URL now points to the correct address, allowing the images to be viewed from any network.